chore(changesets): :package: version packages (#9)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## @edelwud/projen-nextjs@0.1.0

### Minor Changes

-   [`4a8078c`](https://github.com/edelwud/projen-turborepo/commit/4a8078ceabe5acea33c85100cb5ac79e174b809f) Thanks [@MaksimYersh](https://github.com/MaksimYersh)! - inital support nextjs 14