chore(changesets): :package: version packages (#12)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## @yersh/projen-turborepo@0.1.5

### Patch Changes

-   [`10b6feb`](https://github.com/edelwud/projen-turborepo/commit/10b6feb6c111c30cd81897c2cc7f5d742153fa68) Thanks [@MaksimYersh](https://github.com/MaksimYersh)! - include library directory to package