chore(changesets): :package: version packages (#13)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## @yersh/projen-turborepo@0.1.6

### Patch Changes

-   [`4db9030`](https://github.com/edelwud/projen-turborepo/commit/4db90308020252215d6607cc9ea1f477175b7837) Thanks [@MaksimYersh](https://github.com/MaksimYersh)! - move projen JsonFile initialization to constructor