chore(changesets): :package: version packages (#54)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## @yersh/projen-turborepo@0.2.0

### Minor Changes

-   [`fa66d9725f719944b111494942f39cfc85e2568c`](https://github.com/edelwud/projen-turborepo/commit/fa66d9725f719944b111494942f39cfc85e2568c) Thanks [@edelwud](https://github.com/edelwud)! - add `NPM_CONFIG_PROVENANCE` publishing from workflow

## @yersh/projen-nextjs@0.1.2

### Patch Changes

-   [`fa66d9725f719944b111494942f39cfc85e2568c`](https://github.com/edelwud/projen-turborepo/commit/fa66d9725f719944b111494942f39cfc85e2568c) Thanks [@edelwud](https://github.com/edelwud)! - upgraded dependencies