chore(changesets): :package: version packages (#55)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## @yersh/projen-turborepo@0.2.1

### Patch Changes

-   [`ddd07374f323f3835838ad541b85e43accdc9b7e`](https://github.com/edelwud/projen-turborepo/commit/ddd07374f323f3835838ad541b85e43accdc9b7e) Thanks [@edelwud](https://github.com/edelwud)! - configure default command for projen build task