chore(changesets): :package: version packages (#136)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## @yersh/projen-turborepo@0.2.2

### Patch Changes

-   [#135](https://github.com/edelwud/projen-turborepo/pull/135) [`9a421a455a1879b0f908d72fb1eac77146bf8dd7`](https://github.com/edelwud/projen-turborepo/commit/9a421a455a1879b0f908d72fb1eac77146bf8dd7) Thanks [@edelwud](https://github.com/edelwud)! - upgrade dependencies