chore(changesets): :package: version packages (#139)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## @yersh/projen-nextjs@0.2.0

### Minor Changes

-   [#138](https://github.com/edelwud/projen-turborepo/pull/138) [`35574ed`](https://github.com/edelwud/projen-turborepo/commit/35574ed0b120a3887da3bc66764fdd353926e415) Thanks [@edelwud](https://github.com/edelwud)! - upgrade projen package

## @yersh/projen-turborepo@0.3.0

### Minor Changes

-   [#138](https://github.com/edelwud/projen-turborepo/pull/138) [`35574ed`](https://github.com/edelwud/projen-turborepo/commit/35574ed0b120a3887da3bc66764fdd353926e415) Thanks [@edelwud](https://github.com/edelwud)! - upgrade projen package